1. 파일 업로드할 때, backup 디렉토리에 파일 쓰는 방식 변경    - 첫 번째로 업로드한 파일로부터 복사하는 방식으로 변경

+++ src/kr/graha/lib/XMLGenerator.java
@@ -799,21 +799,23 @@ public class XMLGenerator {
 					if(result != null && !result.isEmpty()) {
 						if(result.isArray("_system.filepath")) {
 							List paths = result.getArray("_system.filepath");
-							for(int x = 0; x < paths.size(); x++) {
-								Object path = paths.get(x);
-								if(path != null && path instanceof String) {
-									File dir = new File((String)path);
-									if(dir.exists() && dir.isDirectory()) {
-										File[] ff = dir.listFiles();
-										if(ff != null) {
-											for (File f : ff) {
-												f.delete();
-												if(x == 0) {
-													index++;
+							if(paths != null) {
+								for(int x = 0; x < paths.size(); x++) {
+									Object path = paths.get(x);
+									if(path != null && path instanceof String) {
+										File dir = new File((String)path);
+										if(dir.exists() && dir.isDirectory()) {
+											File[] ff = dir.listFiles();
+											if(ff != null) {
+												for (File f : ff) {
+													f.delete();
+													if(x == 0) {
+														index++;
+													}
 												}
 											}
+											dir.delete();
 										}
-										dir.delete();
 									}
 								}
 							}
